requirements.txt
python-dotenv
flask-socketio
flask
spotipy
pymongo[srv]
